Add entirety of "Chunks of Bytecode" chapter
This is a BLOATED commit, I know.
It's literally my first C code, so there was a LOT of going back and
fourth and checking how I'm supposed to do various things and what the
conventions are in this language.

I feel like at this point, code is acceptable (at least to my low C standards).
I didn't want to artificially split it into separate commits, so here's
the whole thing.